requirements.txt
streamlit
gspread
google-auth
google-auth-oauthlib
matplotlib
pandas
numpy
